Fold PaddleOCR result formatting into recognizer
