+++ test/image/ocr/paddle/test_paddle_series.py
@@ -104,7 +104,7 @@ def test_ocr_image_series_with_paddle_uses_runtime_cache(
             observed_cache_dir_paths.append(cache_dir_path)
 
     monkeypatch.setattr(
-        "scinoephile.image.ocr.paddle._get_runtime_cache_dir_path",
+        "scinoephile.image.ocr.paddle.get_runtime_cache_dir_path",
         lambda *parts: cache_dir_path,
     )
     monkeypatch.setattr(
